updating demoprotocol_schema
