+++ src/App.vue
@@ -518,11 +518,16 @@ export default {
         _.map(eachActivityList, (itemObj) => {
           const newObj = { ...itemObj };
           if (itemObj['@type'] === 'reproschema:Response') {
-            if (itemObj.value instanceof Blob) {
+            if (itemObj.value instanceof Blob && itemObj.mimeType == "audio/wav") {
               const keyStrings = (itemObj.isAbout.split('/'));
               const rId = itemObj['@id'].split('uuid:')[1];
-              jszip.folder(fileName).file(`${keyStrings[keyStrings.length-1]}-${rId}.wav`, itemObj.value);
-              newObj.value = `${keyStrings[keyStrings.length-1]}-${rId}.wav`;
+              jszip.folder(fileName).file(`${keyStrings[keyStrings.length-1]}-${rId}.wav`, itemObj.value); //changed from wav
+              newObj.value = `${keyStrings[keyStrings.length-1]}-${rId}.wav`; //changed from wav
+            } else if (itemObj.value instanceof Blob && itemObj.mimeType == "video/mp4") {
+              const keyStrings = (itemObj.isAbout.split('/'));
+              const rId = itemObj['@id'].split('uuid:')[1];
+              jszip.folder(fileName).file(`${keyStrings[keyStrings.length-1]}-${rId}.mp4`, itemObj.value); //changed from wav
+              newObj.value = `${keyStrings[keyStrings.length-1]}-${rId}.mp4`; //changed from wav
             }
           }
           activityData.push(newObj);
@@ -629,6 +634,7 @@ export default {
       if (!_.isEmpty(this.$store.state.schema)) {
         const order = _.map(this.$store.state.schema['http://schema.repronim.org/order'][0]['@list'],
           u => u['@id']);
+        console.log(order)
         return order;
       }
       return [];
