+++ src/components/Inputs/AudioVideoCheck/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line
+export { default } from './AudioVideoCheck.vue';
