+++ src/components/Survey/Survey.vue
@@ -473,15 +473,19 @@
         _.map(data.response[currentIndex], (itemObj) => {
           const newObj = { ...itemObj };
           if (itemObj['@type'] === 'reproschema:Response') {
-            if (itemObj.value instanceof Blob) {
+            if (itemObj.value instanceof Blob && itemObj.mimeType === "audio/wav") {
               const keyStrings = (itemObj.isAbout.split('/'));
               const rId = itemObj['@id'].split('uuid:')[1];
-              jszip.folder(fileName).file(`${keyStrings[keyStrings.length-1]}-${rId}.wav`, itemObj.value);
-              newObj.value = `${keyStrings[keyStrings.length-1]}-${rId}.wav`;
-            }
+              jszip.folder(fileName).file(`${keyStrings[keyStrings.length-1]}-${rId}.wav`, itemObj.value);//changed from wav
+              newObj.value = `${keyStrings[keyStrings.length-1]}-${rId}.wav`; //changed from wav
+            } else if (itemObj.value instanceof Blob && itemObj.mimeType === "video/mp4") {
+              const keyStrings = (itemObj.isAbout.split('/'));
+              const rId = itemObj['@id'].split('uuid:')[1];
+              jszip.folder(fileName).file(`${keyStrings[keyStrings.length-1]}-${rId}.mp4`, itemObj.value);//changed from wav
+              newObj.value = `${keyStrings[keyStrings.length-1]}-${rId}.mp4`; //changed from wav
           }
           activityData.push(newObj);
-        });
+        }
         // write out the activity files
         jszip.folder(fileName).file(`activity_${currentIndex}.jsonld`,
                 JSON.stringify(activityData, null, 4));
@@ -499,6 +503,7 @@
             this.sendRetry(`${config.backendServer}/submit`, formData);
 
           });
+        });
       },
       async sendRetry(url, formData, retries = 3, backoff = 10000) {
         if (!this.shouldUpload) {
