+++ src/config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  githubSrc: 'https://raw.githubusercontent.com/ReproNim/demo-protocol/main/DemoProtocol/DemoProtocol_schema',
+  githubSrc: 'https://raw.githubusercontent.com/Veronika271/demo-protocol/noAudioVideo/DemoProtocol/DemoProtocol_schema',
   startButton: {
     "en": 'Join',
     "es": 'Participar'
